Move scenario table creation func to conf.py
